Split routes into judge routes and prosecutor routes
