+++ apps/judicial-system/web/src/routes/Prosecutor/CreateDetentionRequest/StepOne/StepOne.spec.tsx
@@ -5,7 +5,7 @@ import userEvent from '@testing-library/user-event'
 import StepOne from './StepOne'
 import { Router } from 'react-router-dom'
 import fetchMock from 'fetch-mock'
-import * as Constants from '../../../utils/constants'
+import * as Constants from '../../../../utils/constants'
 
 describe(`${Constants.DETENTION_REQUESTS_ROUTE}`, () => {
   test('should display an empty form if there is nothing in local storage', async () => {
