+++ apps/judicial-system/web/src/routes/Prosecutor/CreateDetentionRequest/index.ts
