+++ apps/judicial-system/web/src/routes/Prosecutor/DetentionRequests/DetentionRequests.treat.ts
