+++ apps/judicial-system/web/src/routes/Prosecutor/Overview/index.ts
